+++ src/GafferOSL/OSLShader.cpp
@@ -213,7 +213,7 @@ namespace
 boost::container::flat_map<string, string> g_typeOverrides = {
 	{ "vdbVolume", "osl:volume" },
 	{ "dlDisplacement", "osl:displacement" },
-	{ "PxrDisplace", "osl:displacement" }
+	{ "PxrDisplace", "ri:displacement" }
 };
 
 } // namespace